Move public/holidays into public/puzzles
