+++ index.html
@@ -545,7 +545,6 @@
     </main>
   </body>
 
-  <script src="public/holidays.js"></script>
   <script src="public/puzzles.js"></script>
   <script src="public/wordHints.js"></script>
   <script src="public/hint.js"></script>
